+++ src/types/plugins/roomslist/model.d.ts
@@ -2,7 +2,6 @@ export default RoomsListModel;
 declare class RoomsListModel extends Model {
     defaults(): {
         muc_domain: any;
-        nick: any;
         toggle_state: "opened";
         collapsed_domains: any[];
     };
@@ -11,5 +10,5 @@ declare class RoomsListModel extends Model {
      */
     setDomain(jid: string): void;
 }
-import { Model } from "@converse/skeletor";
+import { Model } from '@converse/skeletor';
 //# sourceMappingURL=model.d.ts.map